requirements.txt
flask
mysql-connector-python==9.5.0
secure-smtplib